add blkdat variable lbmont
When HYCOM is run with sshflg=2 the Montgomery potential is offset so
that montg1 in an archive file represents steric SSH.  This makes tidal
analysis easier since non-steric SSH (including tides) is srfhgt-montg1.

The Browning & Kriess barotropic open boundary reads montg1, and it
needs to know if this is the original field or the field corrected via
sshflg=2.  Hence the new blkdat variable lbmont.  This only has an
effect when lbflag=2, and in all legacy cases lbmont will be 0 (false).
Note that lbmont refers to sshflg=2 (true or false) in the outer model run,
and is independent of sshflg in the nested run.

It appears after lbflag in blkdat.input:
   2      'lbflag' = lateral barotropic bndy flag (0=none, 1=port, 2=input)
   0      'lbmont' = baro nesting archives have sshflg=2