major upgrade to latbdt and latbdtf; added lmask_rdnest
The Browning&Kreiss (latbdt) and Flather (latbdtf) barotropic nesting
routines have been overhauled, and an option added to rdnest_in to
mask mean velocity outliers.

A new variable, sports, has been optionally added to ports.input.
It is a scale factor that is applied to all the nested barotropic
velocities.  Use 1.0 (the default if it is absent) to get the
orignal velocities.  If present, it is the 1st line of ports.input:
  1.2   'sports' = scale factor for nested velocity (typically 1.0)
  6     'nports' = number of boundary port sections

Flather uses p-grid nested velocities, [uv]pnst, as in Browning&Kreiss.

The Browning&Kreiss normal transport now always has an along port
1-2-1 smoother applied to it.

Added a compile time parameter, npline, to control how often pline
is updated in latbdt.  It is set to 1 in the provided source code,
but 3 (every third barotropic time step) may reduce noise at the open
boundary.

Added a compile time parameter, ltrans_latbdt, to control the writing
of port transport diagnostics to the model .log file.  It is set to
.false. in the provided source code.

Mean archives can contain large layer velocities near the bottom because
of a mismatch between the p-grid layer thickness mean projected onto cell
faces and u-grid and v-grid layer thickness means.  These no longer occur
for means calculated by the running model, but old in-line means and
off-line means (e.g. monthly or annual) still have this issue.  To prevent
these large velocities effecting the baroclinic nesting, rdnest_in has
a compile time parameter, lmask_rdnest, that, when .true., masks nest
velocities (sets rmunv to zero) when they are outside the near-surface
range.  It is set to .false. in the provided source code.