When coupling, don't call xcstop in HYCOM_Final
When HYCOM runs stand-alone, xcstop calls mpi_finalize on all MPI tasks.
When HYCOM is coupled, xcstop calls mpi_abort because HYCOM may not be
running on all MPI tasks.  This is OK if the xcstop is invoked from an
error condition in HYCOM, but HYCOM_Final is invoked at the end of a
successfull run and it should pass control back to ESMF (say). So the
xcstop is skipped in HYCOM_final.