update the halo of relax.rmu
The relax.rmu file is only read if relax=T.
The rmu array is interpolated from the p-grid to the u and v grids,
and so needs its halo to be populated.  The missing halo update
can cause some points next to the local tile edge to be marked in
error as relaxation-free on the u and/or v grids.  The consequences
are likely minor, but they do cause non-reproducabilty on different
numbers of MPI tasks when relax=T.