Add hybthn to blkdat.input
In hybgen the k-th interface is moved down if layer k is too light,
entraining denser water from layer k+1, and moved up if layer k+1
is too dense, entraining lighter water from layer k.  However,
sometimes layer k is too light and layer k+1 is too dense and we
then need to choose which direction to move the interface.  The
existing approach is to always make the thinner layer thicker.
However, in a Gulf of Mexico case we found that this can lead to one
layer being slightly thinner and very close to the target density while
the other is slightly thicker but very far from its target density.

The new blkdat.input variable, thbthn, tunes the layer choice:

! --- 'hybthn' = HYBGEN: ratio of layer thicknesses to select the thiner
!                 (1.0 for original behaviour, i.e. key only on thickness;
!                  5.0 to favor the larger density anomaly over thickness)

If hybthn is 1.0 the result is as before, but if it is larger than 1.0
then a layer is not identified as "thin" unless it is hybthn times
smaller than the other layer thickness.  If neither layer is "thin" the
layer that is further from its target density is selected and it
entrains water from the other layer.  A value of 5.0 worked well in the
Gulf of Mexico, but this approach has not been tested elsewhere.

All blkdat.input files need to be updated with hybthn:

   0.01   'hybiso' = HYBGEN: Use PCM if layer is within hybiso of target density
   1.0    'hybthn' = HYBGEN: ratio of layer thicknesses to select the thiner
   3      'hybmap' = HYBGEN: remapper  flag (0=PCM, 1=PLM,  2=PPM, 3=WENO-like)

Existing results are repeatable provided hybthn=1.0.