Added the macro MOMTUM_CFL
The compile time macro MOMTUM_CFL adds a CFL limiter to momtum, similar
to the existing MOMTUM4_CFL in momtum4.  In both cases velocities are
clipped if they exceed a CFL limit, and if the clipping exceeds 0.5 m/s
at a location a diagnostic is written to the .log file.

This is a macro because CFL should not be exceeded when HYCOM runs
with appropriate parameters and because clipping is not a physical
process.  If MOMTUM_CFL is applied at compile time, the model .log
file should be monitored for "velocty clipped" warning messages.

The fatal stop for large negative dp's in loop 19 of cnuity has been
removed, because these might be corrected in loop 15.