Added the macro MASSLESS_1MM
In mxkprf.f the "lowest substantial mass-containing layer" is at
least 10cm thick by default.  The compile time macro MASSLESS_1MM
changes this to 1mm, which is the minimum possible in this context.

Note that in hybgen, the "lowest substantial mass-containing layer"
is at least 0.001mm thick.  This difference is due to avoiding ill
conditioned tridiagonal matrices in mxkprf.f

This is implemented as a compile time macro because 10cm has worked
well in most cases and 1mm has not be extensively tested.  The 10cm
criteria has lead to anomalous T and S values in GOFS 3.1 when a
layer is always less than 10cm thick at the bottom.  So setting
MASSLESS_1MM is advisable when running with data assimilation, and
should arguably be the default.

Make.csh has been updated to document the available OCN_MISC flags.
It also now defaults to -DEOS_SIG2 and -DEOS_17T.