Better support for doubly periodic domains
Doubly periodic domains must be on a f-plane and are primarily used for
1-D (single column) cases, with idm=jdm=5 and no variation in x or y,
but also for some idealized test cases such as an isolated gyre.

These domains were only previously supported on a single task, but
now multple tasks (i.e. MPI) is now allowed provided there is only
one tile in the j direction (mpe=1).