Add diagnostic tracers
Diagnostic tracers use the existing dynamic tracer infrastructure, but
since they are diagnostic they are not written to the rerstart file.

Supported by two new blkdat variables mtracr and dtrflg, but if mtracr
is 0 dtrflg is not included:

   0      'ntracr' = number of tracers (0=none,negative to initialize)
   0      'trcflg' = tracer flags      (one digit per tr, most sig. replicated)
   3      'mtracr' = number of diagnostic tracers
 801      'dtrflg' = diagnostic tracer flag (one per tracer, 801-899)
 802      'dtrflg' = diagnostic tracer flag (one per tracer, 801-899)
 803      'dtrflg' = diagnostic tracer flag (one per tracer, 801-899)

At present only three diagnostic tracers are provided:

  dtrflg=801: change in layer thickness   due to hybgen, m/day
  dtrflg=802: change in layer temperature due to hybgen, degC/day
  dtrflg=803: change in layer salinity    due to hybgen, psu/day

See hybgen.F90 for how these are calculated.  Tracer flags 804 to 899
are available to implement other diagnostic tracers in the future.
The only restrictions are that they are nominally layer averaged
quantities, and they must be calculatable every time step.  Like
visc and diff, they could also be layer interface quantities but
then their means would not be accuate.