Bugfix for epmass=1 when oneta is small
The actual depth is oneta * depth, with oneta limted to be at least oneta0.
There was a bug in epmass=1 for oneta close to oneta0.  This has been fixed
by:
  a) replacing oneta limit (oneta0), internally with pbot limit (pbotmin)
  b) allowing evaporation to extend below a thin enough layer 1
  c) clipping evaporation if it would drive oneta below oneta0

Note that SSH is typically greater than -2 m without tides and so a
small oneta is unlikely to occur for a typical minimum rest depth unless
tides are included.

The reporting of oneta=oneta0 was every 3 time steps, but now it is when
other extended diagnostics are reported (e.g. every diagfq days).