Add ocnscl for finer control of relative winds
Existing options are wndflg 4,5 to use relative wind U10-Uocn;
and -4,-5 to use absolute wind U10.  The new blkdat variable ocnscl
sets wind to U10-ocnscl*Uocn.  So 0.0 is absolute winds and 1.0
is full relative winds.  Setting ocnscl to 0.7 is a popular choice.

This adds a line to all blkdat.input files, for example:

  -5      'wndflg' = wind stress input flag (0=none,1=uv-g,2,3=p-grd,4,5=wnd10m)
   0.0    'ocnscl' = scale factor for Uocn in relative wind (0.0: absolute wind)