Add hybthk to expand thin layers
HYBGEN often creates a thick-thin-thick layer structure.  Which
reduces the effective vertical resolution.

The bkldat variable hybthk controls expanding such thin layers with
water from adjacent layers above and below.

All blkdat.input files must be updated.  Setting this to 0.0 recovers
the original behaviour.  The maximum value is 0.9 and this is often
a good choice.  For example:

0.01 'hybiso' = HYBGEN: Use PCM if layer is within hybiso of target density
5.0  'hybthn' = HYBGEN: ratio of layer thicknesses to select the thiner
0.9  'hybthk' = HYBGEN: thick-thin-thick ratio to expand the thin layer
3    'hybmap' = HYBGEN: remapper flag (0=PCM, 1=PLM, 2=PPM, 3=WENO-like)

Note that hybthk cannot distinguish between pathological thick-thin-thick
layers and layers that are that way because of dynamics, e.g. in strong
fronts and eddies.  So use with care.