Add surface tracers
Surface tracers are 2-D fields that are added at the end of the surface
archive files.  Note that "surface" fields are not necessarily at the
surface, for example u_btrop and v_btrop can be in standard archs files.

Supported by two new blkdat variables mstrcr and strflg, but if mstrcr
is 0 strflg is not included:

   0      'mtracr' = number of diagnostic tracers
   1      'mstrcr' = number of surface    tracers
 703      'strflg' = surface tracer flag (70X: 3=oneta,4=pbavg)

At present only four surface tracers are provided:

701: disp_ld - dissipation from linear    drag, time mean
702: disp_qd - dissipation from quadratic drag, time mean
703: oneta   - 1+eta
704: pbavg   - bottom pressure

Tracer flags 705 to 799 are available to implement other surface tracers
in the future.  If the fields are already available, they can be added
in mod_archive.F90.  Otherwise they can be calculated anywhere in the
model's time step and stored in stracr(:,:,ktr), and also added to
mod_archive.F90 if you want a non-generic name for the tracer.

The HYCOM-tools utilities hycomproc, archv2data2d and archv2ncdf2d now
support these 2-D tracer fields in surface archive (archs) files.