Add cbtidc to include tidal velocities in bottom speed
If cbtidc is non zero, tidal.obs[uv]ReIm.[ab] contains eight tidal
velocity consituents with cbtidc controling which consituents are on.
Tidal velocities are added to the near bottom velocities that
determine the total quadratic bottom drag coefficient.

In addition, the effect of cbar has been changed.  A positive cbar
is still a linear drag with coefficient cbar*cb as before, but now
if cbar is negative cbar.[ab] is input and this spatially varying
speed is included in the calculation of bottom speed.  This is a
better parameterization of the effect of tides on bottom drag than
treating it as a linear drag. The usual cbar field is RMS tidal
speed, which is 0.707 * tidal Amplitude.

There is one new blkdat.input variable, for example:

   0.0    'cbar  ' = rms flow speed (m/s) for linear bottom friction (-ve input)
  -2.5e-3 'cb    ' = coefficient of quadratic bottom friction (-ve input)
00000001  'cbtidc' = tidal velocities: 1 digit per constituent (Q1K2P1N2O1K1S2M2)

It is an error for cbtidc to be non-zero when tidflg is positive, i.e.
when there is tidal forcing.  It is not an error for cbar to be non-zero
when tides are active (tidflg>0 or cbtidc positive) but this is not
recommended.

There are three new surface tracers.  A typical cbtidc run might include:

   4      'mstrcr' = number of surface    tracers
 704      'strflg' = surface tracer flag (70X: 04=pbavg)
 721      'strflg' = surface tracer flag (70X: 21=efold_cb)
 722      'strflg' = surface tracer flag (70X: 22=spdbot)
 723      'strflg' = surface tracer flag (70X: 23=spdtid)

The spdtid tracer is only allowed if cbtidc is not zero, since it
writes out the tidal speed.  The efold_cb and spdbot tracers can
be used to diagnose bottom friction in any setup.  The pbavg tracer
is not strictly required, since pbavg=srfht-montg, but is now
often on for convenience.