printout now ok for kdm<1000 and idm,jdm<100,000
All large arrays are allocated at run time, and so HYCOM has no upper
limit on idm, jdm and kdm.  However, many print statements used i2, or
i3 with the expection that the 3rd digit is a space, for kdm and 2i4,
or 2i5 with the expection that the 3rd digit is a space, for idm,jdm.

All such print statements have been updated by adding a digit (i3,i4,
2i5,2i6) and so kdm of 1 to 999 and idm or jdm of 1 to 99999 are
fully supported.  There are extra spaces in .b archive files and in
the log output.