Bugfix for sshflg=2 or 3
Setting sshflg to 2 or 3 reads in mean Montgomery Potential and SSH from
relax.montg.[ab] and adds the correction montg_c = (SSH.mn - MONTG.mn)
to the reference bottom pressure.  This makes the layer 1 Montgomery
Potential the Baroclinic SSH, which is a better approximation than
Steric SSH for looking at internal tides.

Commit cb17d03 introduced sshflg = 3 to calculate both Steric SSH and
Baroclinic SSH (Montgomery Potential), but it introduced a bug in the
sign of montg_c that made the Montgomery Potential inaccuarate.  This
bug is fixed here.