Add checksum option to mod_pipe
The mod_pipe package was originally designed to connect a master and
slave HYCOM run, to compare all prognostic fields after each major
subroutine call, and stop if they differ.

The connection can either be in a single hycom executable compiled with
the OCEANS2 macro, or betwen two executables on the same node
communicating via a named pipe.  Neither option is easy to setup and
they can't easily be used when coupled to other components like sea ice.

As as alternative, with the PIPE_CHECK option there is no slave and the
master writes out a 32-bit MurmurHash3 checksum for each field referenced.
Its primary use is to write checksums from two runs that are known to
differ (e.g. restarts not the same) when they should be identical.  A
difference of the output checksums will indicate which fields first
differ and in what subroutine.  A master slave setup would idenitify
the i,j location and size of the difference, but knowing which field
differs is often enough.  This option runs very slowly, and it does
not stop on a difference (since it isn't looking for differences), so
the runs should typically be short (a few model hours).